move connection code for TCP and connector to separate files
