+++ cloud-sql/mysql/servlet/src/main/java/com/example/cloudsql/ConnectorConnectionPoolFactory.java
@@ -1,0 +1,87 @@
+/*
+ * Copyright 2018 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.example.cloudsql;
+
+import com.zaxxer.hikari.HikariConfig;
+import com.zaxxer.hikari.HikariDataSource;
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import javax.sql.DataSource;
+
+
+public class ConnectorConnectionPoolFactory extends ConnectionPoolFactory {
+
+  // Saving credentials in environment variables is convenient, but not secure - consider a more
+  // secure solution such as https://cloud.google.com/kms/ to help keep secrets safe.
+  private static final String INSTANCE_CONNECTION_NAME =
+      System.getenv("INSTANCE_CONNECTION_NAME");
+  private static final String INSTANCE_UNIX_SOCKET = System.getenv("INSTANCE_UNIX_SOCKET");
+  private static final String DB_USER = System.getenv("DB_USER");
+  private static final String DB_PASS = System.getenv("DB_PASS");
+  private static final String DB_NAME = System.getenv("DB_NAME");
+
+  @SuppressFBWarnings(
+      value = "USBR_UNNECESSARY_STORE_BEFORE_RETURN",
+      justification = "Necessary for sample region tag.")
+  public static DataSource createConnectionPool() {
+    // [START cloud_sql_mysql_servlet_connect_connector]
+    // [START cloud_sql_mysql_servlet_connect_unix]
+    // The configuration object specifies behaviors for the connection pool.
+    HikariConfig config = new HikariConfig();
+
+    // The following URL is equivalent to setting the config options below:
+    // jdbc:mysql:///<DB_NAME>?cloudSqlInstance=<INSTANCE_CONNECTION_NAME>&
+    // socketFactory=com.google.cloud.sql.mysql.SocketFactory&user=<DB_USER>&password=<DB_PASS>
+    // See the link below for more info on building a JDBC URL for the Cloud SQL JDBC Socket Factory
+    // https://github.com/GoogleCloudPlatform/cloud-sql-jdbc-socket-factory#creating-the-jdbc-url
+
+    // Configure which instance and what database user to connect with.
+    config.setJdbcUrl(String.format("jdbc:mysql:///%s", DB_NAME));
+    config.setUsername(DB_USER); // e.g. "root", "mysql"
+    config.setPassword(DB_PASS); // e.g. "my-password"
+
+    config.addDataSourceProperty("socketFactory", "com.google.cloud.sql.mysql.SocketFactory");
+    config.addDataSourceProperty("cloudSqlInstance", INSTANCE_CONNECTION_NAME);
+
+    // [END cloud_sql_mysql_servlet_connect_connector]
+    // Unix sockets are not natively supported in Java, so it is necessary to use the Cloud SQL JDBC
+    // Socket Factory to connect.
+    // Note: For Java users, the Cloud SQL JDBC Socket Factory can provide authenticated connections
+    // which is preferred to using the Cloud SQL Proxy with Unix sockets.
+    // See https://github.com/GoogleCloudPlatform/cloud-sql-jdbc-socket-factory for details.
+    if (INSTANCE_UNIX_SOCKET != null) {
+      config.addDataSourceProperty("unixSocketPath", INSTANCE_UNIX_SOCKET);
+    }
+    // [START cloud_sql_mysql_servlet_connect_connector]
+
+    // The ipTypes argument can be used to specify a comma delimited list of preferred IP types
+    // for connecting to a Cloud SQL instance. The argument ipTypes=PRIVATE will force the
+    // SocketFactory to connect with an instance's associated private IP.
+    config.addDataSourceProperty("ipTypes", "PUBLIC,PRIVATE");
+
+    // ... Specify additional connection properties here.
+    // [START_EXCLUDE]
+    configureConnectionPool(config);
+    // [END_EXCLUDE]
+
+    // Initialize the connection pool using the configuration object.
+    DataSource pool = new HikariDataSource(config);
+    // [END cloud_sql_mysql_servlet_connect_connector]
+    // [END cloud_sql_mysql_servlet_connect_unix]
+    return pool;
+  }
+
+}
